feat: refactor loop PIP to separate cocmponent
