+++ domains/DataCollection/Forms/index.js
@@ -21,7 +21,7 @@ const Forms = (props) => {
     pinnedForms,
     surveyingUser, surveyingOrganization,
     surveyee, setSurveyee,
-    customForm, setCustomForm, navigateToRoot
+    customForm, navigateToRoot
   } = props;
 
   const [consent, setConsent] = useState(false);
@@ -60,7 +60,6 @@ const Forms = (props) => {
                 surveyingUser={surveyingUser}
                 surveyingOrganization={surveyingOrganization}
                 customForm={customForm}
-                setCustomForm={setCustomForm}
               />
             </View>
           </TouchableWithoutFeedback>
